+++ CHANGELOG.md
@@ -5,7 +5,7 @@
     import and export databases, for each driver, with placeholders;
 * `__exportExecutableWithCompression()` and `_importExecutableWithCompression()`
     methods provided by the `Driver` class have been removed and incorporated
-    into `_exportExecutable()` and `_importExecutable()`;
+    into the new `_getExportExecutable()` and `_getImportExecutable()`;
 * `BackupTrait::$validExtensions` has been removed and replaced by the
     `DATABASE_BACKUP_EXTENSIONS` constant;
 * postgres and sqlite commands are also properly escaped;